terraform/OCI: Fix extra volume device names
The device attachment resource started to reuse device names after
more than 10 devices were created. This happens when there are
several instances and more than a few devices per instance.

The original intent was to give each instance volumes named
"vdb", "vdc", "vdd", and so on. To accomplish that, adopt the
same mechanism I worked out for Azure's manage_disk LUNs.

The "required_providers" block is removed from provider.tf because
it created a provider alias that the new module failed to recognize.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>