scripts/bringup_guestfs.sh: fix silly directory permission fix
The bringup process for libvirt URI system support (not session,
so all debian based distros) requierss us to be paranoid about the
permissions of our storage directory where we place our libvirt
storage pool, and guestfs images. That's no reason to be hammering
with a sledge hammer a crazy sudo chown -R on a target storage path.
Specially when that directory may be shared with other users.
Essentially this now does:

sudo chown -R foo.libvirt-qemu /path/to/storage/

For small systems that may be OK, but that's nuts for partitions
which are shared with other users, which is the way that kdevops
storage pools was designed to support in the first place. Just remove
this as this is very dangerous. I will soon post a replacement with
an ansible check and it just checks for the permissions and complains
and fails if we don't have the right permissions.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>