provisioning: Add a "make status" target
If you've walked away from the keyboard and forgotten whether there
are target nodes running or waiting for the next step, it's easy
to "make mrproper" and wipe out the ssh keys and other state that
Ansible and terraform use to access those nodes.
With guestfs, it's as easy as "virsh list" to see what's running.
But with terraform, the nodes are running in another geographic
location entirely.

Add a make target that can probe to see if we've left something
cooking on the back burner.

Before "make bringup" or after "make destroy":

$ make status
make: Nothing to be done for 'status'.
$

After a "make bringup with guestfs:

$ make status
 Id   Name        State
-------------------------
  2   test-nfsd   running
$

After a "make bringup" with terraform:

$ make status
There are 4 active terraform resources on 'gce'.
{
  "gce-nfsd" = "35.243.237.209"
}
$

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>