terraform/GCE: Add a "Resource Location" submenu
Introduce the strategy of moving these submenus to their own separate
Kconfig files to eventually enable their generation by querying the
provider with a script (similar to "make refs-default"). This would
allow the Kconfig menu content to be adjusted based on what is
available on the provider platform and what is accessible to the
project.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>