terraform/GCE: Enable addition of new regions
Set up a choice menu of regions. This might be eventually expanded
to include more regions, or it might be populated via script, based
on the project's access and the availability of new regions on
Google Cloud.

The purpose is to provide users with a small set of valid selections
rather than the blank slate of a free-form text field.

Note: the Kconfig symbols are named with the concatenated ISO 3166
code for brevity. See:

https://gist.github.com/rpkim/084046e02fd8c452ba6ddef3a61d5d59

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>