terraform/OCI: Add a "Resource location" submenu
Introduce the strategy of moving these submenus to their own
separate Kconfig files to eventually enable their generation by
querying the provider with a script (similar to "make
refs-default"). This would allow the Kconfig menu content to be
adjusted based on the intersection of what is currently available
on the provider platform and what is accessible to the tenant.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Reviewed-by: Chandan Babu R <chandanbabu@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>