terraform/OCI: Add an availability domain selector
In order to balance load, OCI adds a prefix to availability
domain names. This prefix is fixed but unique for each user.

The user has to look up the prefixed name (at least once) before
specifying an availability domain name for terraform, which is
not user-friendly.

Fortunately, terraform provides a data source to look up the
user's unique prefix and form the correct full domain name for
terraform. So the Kconfig selector only needs to pick an integer
from 1 to 3, and that hides a needless detail from our user.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Reviewed-by: Chandan Babu R <chandanbabu@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>