terraform/OCI: Simplify image selection
Currently OCI requires users to hunt down the OCID of the
particular OS image they wish to deploy by walking through a list of
images and regions on Oracle's web site, then copying and pasting
the OCID from a region list. This is not terribly user friendly.
I'd like to provide a simple menu to make it easier to select an
image without spending an hour doing research.

The terraform oracle/oci provider has an "oci_core_images" data
source that invokes the ListImages API to search for the OCID of an
OS image based on it's family and name. Neat!

Unfortunately, according to:

  https://github.com/oracle/terraform-provider-oci/issues/352

the ListImages API returns a list that is dynamically generated on
every call. As images are added and removed from OCI, that list will
change over time -- in fact it can change between two "terraform
apply" calls, which might cause a "terraform destroy" to fail. We
want kdevops to have reliable bringup and shutdown to enable
automation.

As described in:

https://docs.public.oneportal.content.oci.oraclecloud.com/en-us/iaas/Content/dev/terraform/ref-images.htm#top

the best practice is to stuff fixed image OCIDs into a terraform map
variable. It's ugly and requires upkeep, but it guarantees the OCIDs
are stable from one "apply" to the next.

kdevops now deals with more than a handful of regions and operating
systems, so instead of a terraform map variable, I'm going with a
solution where Kconfig selects an image and sets its OCID in the
gen_tfvars template. That is functionally equivalent to a map
variable but IMO it will be easier for us to maintain (either by
hand or automatically via a script).

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Reviewed-by: Chandan Babu R <chandanbabu@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>