devconfig: add gitconfig /mirror/ sanity check
When you leverage the use of /mirror/ say as an NFS server then your
system will use a git service to allow guests with libvirt to connect
to the NFS server through your own host. This works quite well, however
git will prevent access if the directory is not owned by you.

To fix this the *host*, ie, localhost, must trust the directories,
without this will fail on both the host and the guests:

/usr/bin/git ls-remote git://192.168.122.1/mirror/fstests.git -h refs/heads/HEAD

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>