terraform/AWS: One default value to rule them
There are 3 sets of default values for each terraform setting.

1. The default value set in terraform/oci/Kconfig
2. The default value in playbooks/roles/gen_tfvars/defaults/main.yml
3. The default value specified in terraform/oci/vars.tf

Often these defaults are different values. It's confusing for
developers, but worse is that it can result in hard-to-diagnose
terraform provider error messages (eg, "Where did that 'invalid'
come from?").

IMO a better plan is to let Kconfig provide the only default value,
and remove the other two. We then have a single authoritative
default value for each setting.

If a required variable is missing a value during "terraform plan",
that will always be a local bug, not something exposed to the
provider to report via an obscure error message.

There are two exceptions:

- For  Kconfig booleans, often the variable is defined with a "y"
  value in the "true" case, but not defined at all in the "false"
  case. So for booleans, the gen_tfvars template needs a default
  value in gen_tfvars/defaults/main.yml of "false".

- Where a terraform resource expects an argument to be not present
  in some configurations, then vars.tf needs to explicitly define a
  default value of "null". Only in those cases it is OK for no
  value to be provided by Kconfig.

This is a refactoring change; no behavior change is intended.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>