terraform/Azure: Rework identity settings
I noticed that Buildbot logs contain a full listing of environment
variables, and that exposes Azure authentication material handled
via these variables if the logs are visible publicly.

So, instead, let's go with storing the secrets in a dot-directory,
separate from the kdevops .config file, which might be managed
via a source control system. Best to keep secrets out of those.

Suggested-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>