gen_node: add debugger support for guestfs libvirt guests
The motivation behind enabling the support for the debugger is to
make kdevops user friendly for bug fixing and troubleshooting, thereby
saving the time needed to configure guest during bringup. When enabled,
libvirt_gdb_baseport is assigned with a base port number extracted
from the md5sum of the script get_gdb_base_port.sh to which we add an index
for every guest during gen_nodes bringup. The bringup is also modified
to verify that no port conflict exists and if does, will fail informing
the user to change the base port.

This feature is disabled by default as kdevops is heavily used
by CI where this support is not needed. However for folks relying
on kdevops for kernel development and testing, can now leverage this
option to help with debugging.

Signed-off-by: Swarna Prabhu <s.prabhu@samsung.com>
[mcgrof: added libvirt_enable_gdb checks in a few places]
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>