Kconfig: Convert the 9p option to a choice menu
I'm about to add a third method for building the kernel that
excludes the use of 9p. As a pre-requisite to that change, wrap the
two existing options with a choice menu, since they are exclusive
of each other.

This introduces a boolean for non-9p build configurations, to be
used in subsequent patches.

No behavior change is expected.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>