bootlinux: Add a new builder choice
Currently, for cloud configurations or when 9p is disabled, kdevops
builds the test kernel on each test runner. This is inefficient,
and gets worse as the test matrix for a single kernel version scales
out.

Instead we want to build the test kernel once and make those build
artifacts available for test runners to install. This would be
similar to what KOTD does now, except it does not require setting up
a separate yum repo.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>