terraform: Remove local ssh config first
"terraform destroy" can fail for a host of reasons. Because the step
that resets the local ssh config follows the "destroy" step, it does
not get run if "destroy" fails. That leaves the ssh config in a
potentially broken state when "terraform destroy" doesn't work.

Switch the order of the tasks so the local ssh config is always
cleaned up by "make destroy".

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>