guestfs: Convert scripts/bringup_guestfs.sh to Ansible
The bringup_guestfs.sh script is converted to Ansible to enable
multiple instances of virt-builder to run in parallel, making guest
bring-up faster. This also gives us a fighting chance to integrate
terraform and guestfs ssh key management in the future.

In addition to those benefits, the new Ansible code takes a stab at
being more idempotent than the shell script was. Idempotency is
achieved by checking that each guest is defined, and then defining
it if not; then using community.libvirt.virt to boot the guest,
which does nothing if the guest is already running.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>