guestfs: "make destroy" can be more surgical
The libvirt.virt module isn't idempotent. To work around that, I
simply set "failed_when: false" to ignore cases where a target node
was already shut down or undefined.

But that means "make destroy" ignores real failures.

Instead of squelching all failures when shutting down and undefining
target node, make a list of nodes in each state, and compare it
against the inventory. Only take action when there is something to
do. And, mark the target node as changed when that action is taken.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>