workflows: Add a kconfig setting for installing kernels via package
When the "Target Linux kernel to use" is set to "Upstream Linux or
custom kernel", the bootlinux role is enabled in order to download,
build and install the kernel to be tested.

When installing the kernel from .rpm or .deb packages, however, the
download and build phases have already been done, so it doesn't make
sense to add an option under the "Get and install Linux from git"
menu to make that happen.

Instead, add an option directly to the "Target Linux kernel to use"
menu that will enable the ability to install the test kernel from
packages.

For the moment, this option does nothing.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>