bootlinux: Clean up a grub set-up task
The "target_kernel" variable seems to be unused, so remove it.

Address several ansible-lint complaints:
 - command-instead-of-shell
 - fqcn[action-core]: Use FQCN for builtin module actions (shell).
 - jinja[spacing]: Jinja2 spacing could be improved
 - name[template]: Jinja templates should only be at the end of 'name'
 - no-changed-when
 - schema[tasks]: 'become_method' must be one of the currently available values
 - yaml[line-length]: Line too long (176 > 160 characters)
 - yaml[truthy]: Truthy value should be one of [false, true]
 - yaml[brackets]: Too many spaces inside brackets
 - fqcn[action-core]: Use FQCN for builtin module actions (set_fact).

Note: The preferred way to select a shell script and run it on a
managed system is to use the template module to copy the selected
script to each target host, and then use the Ansible scripts module
to run it. That could be for another day.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>