devconfig: Boost update speed
On RedHat-based systems, increase the number of concurrent downloads
to reduce the amount of time it takes to download and install
packages.

I've set the maximum number of concurrent downloads to 10, which is
more than the default, but not too high that kdevops test systems
will cause a flood of work that targets community DNF repos. If
necessary, this value can be made into a Kconfig option, but let's
leave that until we actually see a need for it.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>