steady-state: fix identation on Makefile
ChatGPT Codex generated this code, it gets a B rating as I only found
one bug and this was it. I told it to borrow the stead-state code from
my plot-sysbench code [0] and also take my fio-tests language and
Kconfig logic from fio-tests [1].

And so I'll be folding in fio-tests to kdevops soon using kdevops as
this is proving to be quite simple with ChatGPT codex. I can say then
with more confidence that the architecture behind kdevops is making
adding more workflows with things like ChatGPT codex super simple. We
should be able to scale faster by now and recent developments / tools.

Link: https://github.com/mcgrof/plot-sysbench # [0]
Link: https://github.com/mcgrof/fio-tests # [1]
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>