steady-state: quite a bit of fixes
The original implementation forgot to just open code the
pre-fill still as an ansible task. Claude AI did this for me,
and with a bit more love I dealt with the odd ansible oddities
found. Now that we have an open coded ansible version of the
optimal prefill step with kdevops integration just remove the
script.

Clarify the heuristics of the optimal prefill step on the
SSD_STEADY_STATE_PREFILL_BLOCKSIZE documentation.

Let's also ensure we provide steady state output files, which is
useful to visualize the steady state process for bw and iops.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>