docs: declare kdevops a software 3.0 enabler
For a long time we've worked hard to make it easier to extend kdevops
with strict rules and clear declarative language. We've been so strict
on some of these policies I'm quite sure it discouraged some developers.
The projections and hope was that these strict goals would reduce the
barrier to the use of generative AI for extending kdevops.

Although we've been experimenting with different AI agents on kdevops
for a while now, advanced AI agent tools such as Claude Code bring this
closer to home, enabling prompts to complete about 70-95% of the tasks.
In a short span of about one week we've been able to cut through a huge
back log of work which would have otherwise taken us considerable amount
of time. This, and our automated tests put in place gives us confidence
to declare that the time is here and now and it should be easy for
others to also extend kdevops for their needs with generative AI.

Extend our documentation with the rationale of the implicit long term
goals we had, make these explicit, and also provide guidance for Claude
Code. We also provide some recent prompt examples and respective commits
merged which should be useful for users wishing to look to extend
kdevops.

CLAUDE.md [0] is a special file that Claude automatically pulls into
context when starting a conversation. This makes it an ideal place
for documenting:

  - Common commands folks should be aware of that we run
  - Core files and utility functions
  - Code style guidelines
  - Testing instructions
  - Repository etiquette (e.g., branch naming, merge vs. rebase, etc.)
  - Developer environment setup
  - Any unexpected behaviors or warnings particular to the project
  - Other information you want Claude to remember

We extend this to ensure the DCO is respected and also give it
a few prompts examples and target historical commits with respective
grades.

Also, extend the video documentation links with the recent kdevops talk
in Open Source Summit North America [1] where I mentioned for the first time
publicly that kdevops was Software 3.0 enabler for Linux kernel development.

Link: https://www.anthropic.com/engineering/claude-code-best-practices # [0]
Link: https://www.youtube.com/watch?v=VF-jr_ZE-9Y&list=PLjaT52x3HVboZtdwZnONSHQHM8217ELcc # [1]
Acked-by: Chuck Lever <chuck.lever@oracle.com>
Acked-by: Daniel Gomez <da.gomez@samsung.com>
Tested-by: Daniel Gomez <da.gomez@samsung.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>