bootlinux: fix making 9p default if using libvirt
Commit a5092b07fe ("Kconfig: Convert the 9p option to a choice menu")
enhanced the semantics to use a choice instead of just bools to
determine how we are going to build the kernel. But it ended up making
the target option to build the kernel and clone it the default even if
you are using libvirt. Fix this so that we only select the target
if we're not using libvirt. This upkeeps previous behaviour and
expectations.

Fixes: a5092b07fe ("Kconfig: Convert the 9p option to a choice menu")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>