bootlinux: add support for custom refs on dev kernels on the CLI
We already have support to easily test arbitrary kernels and refs by
just using the LINUX_TREE and LINUX_REF environment variables. We
leverage this for our CIs:

  * LINUX_TREE: the target tree     for baseline group (A)
  * LINUX_REF:  the target tree ref the baseline group (A)

Now that we have AB-testing support we want to first add support so that
a simple TEST_AB=y will enable KDEVOPS_BASELINE_AND_DEV and then we want
to be able to customize the target dev tree and ref. We do this with two
other environment variables:

  * LINUX_DEV_TREE: the target tree     for dev group (B)
  * LINUX_DEV_REF:  the target tree ref the dev group (B)

You just need to make sure to also pass TEST_AB=y.

By leveraging this we can easily use AB testing in the future on CIs
to compare and contrast different kernel releases against any target
workload we have.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>