reboot-limit: simplify what gets selected
If you try to take the reboot-limit workflow for a spin right now
you will run into a conflict with networkd.service and the bringup
we have on debian-testing. I have narrowed this down to the few extra
selects we do, and the reason is that we have not seen these issues
before on our CI. Testing with these disabled fixes the issue so just
disable those selects as we already do enough of a good job of enabling
sensible defaults.

Reviewed-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>