linux: use stable ref for baseline and dev for dev group
Right now we have the default tree being selected for the "baseline"
group. However when booting with BOOTLINUX_AB_DIFFERENT_REF we end up
with a baseline group and "dev" group.

Is is therefore odd, to end up up with the "stable" kernel on the "dev"
group, and the latest kernel on the "baseline" group. Invert this logic,
so the dev group always has the latest and greatest to match expectations
that dev groups are the bleeding edge.

Now we get more sensible defaults:

$ make defconfig-mmtests-ab-testing-thpcompact KDEVOPS_HOSTS_PREFIX="lpc"
$ make
$ make bringup
$ make bringup
$ make uname

lpc-thpcompact-dev               6.17.0-rc4-00047-ge6b9dce0aeeb
lpc-thpcompact                   6.15.0

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>