terraform/aws: Add a script to generate the Kconfig.locations menu
Enable kdevops to quickly pick up new regions and availability zones
whenever they are introduced by AWS. The regions that are discovered
and output by this script are those that the user has opted-in or are
"no opt-in required".

boto3 reuses connections to the API endpoint, making the generation
of the Kconfig menu quick. Jinja2 lets us manage a large amount of
bespoke Kconfig help text painlessly. The new script reports some
basic authentication issues and has a "raw" output to enable
troubleshooting.

This implementation is still missing geographic location information
that is retrieved via the 'pricing' API. IMO that can be introduced
later.

To use:

  $ terraform/aws/scripts/gen_kconfig_location >
	terraform/aws/kconfigs/Kconfig.location

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>