devconfig: drop locales-all leftovers
When locales-all support was added in commit 2514a694 ("guestfs: fix
locale warnings in VM deployment"), we added for all distros in the
virt-builder.j2 template as well as Ansible tasks to ensure idempotency
on previous base images.

But we just need to ensure en_US.UTF-8 is available on the guests
image, so keep it this config in both template and tasks but remove
locales-all.

While we are at it, also move the locale configuration after all
deps are installed and to the distro agnostic as we do it anyways in
virt-builder.j2 to ensure idempotency.

Fixes: 6bb3bfb2 ("base_image: Skip "locales-all" package installation")
Signed-off-by: Daniel Gomez <da.gomez@samsung.com>