devconfig: drop locales-all leftovers
When locales-all support was added in commit 2514a694 ("guestfs: fix
locale warnings in VM deployment"), it was added for all distros in the
virt-builder.j2 template as well as Ansible tasks to ensure idempotency
on previous base images.

But we just need to ensure en_US.UTF-8 is available on the guests
image, so keep just this config in both template and tasks but remove
locales-all.

While we are at it, also move the locale configuration after all deps
are installed and to the distro agnostic task section. We do it anyways
for all distros in virt-builder.j2, this ensures idempotency in all
of them.

Fixes: 6bb3bfb2 ("base_image: Skip "locales-all" package installation")
Link: https://github.com/linux-kdevops/kdevops/actions/runs/18347971703/job/52259752236 [1]
Link: https://lore.kernel.org/r/20250929-locales-all-removal-v1-1-4177411d3e44@samsung.com
Signed-off-by: Daniel Gomez <da.gomez@samsung.com>