pynfs: Fix the xdrlib dependency for Debian
When running "make pynfs" in the current directory, I get to the
"Install pynfs build dependencies" task, which fails with the
message "No package matching 'python3-standard-xdrlib' is
available". The test system is running Debian 11.

Python 3.13 removed xdrlib (PEP 594), and we recently adjusted
install-deps/redhat/main.yml to accommodate that.

I found that the xdrlib module is already part of Python 3.9's
standard library on Debian 11. The package python3-standard-xdrlib
doesn't exist on Debian 11 because xdrlib is included by default.

Remove the python3-standard-xdrlib package and adopt the solution
used by install-deps/redhat/main.yml, which is to pull in xdrlib3
when xdrlib is not included in the standard Python library (this
should be only on Debian 13).

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>