bootlinux: Ensure nfsd host gets an updated Linux kernel
Commit 434df33309e7 ("workflows: linux: remove explicit inventory
in ansible wrappers") added "--limit 'baseline:dev' to the bootlinux
Makefile invocation to exclude localhost from the list of hosts that
get a kernel update.

This, however, excludes the nfsd host. When it is part of a test
workflow, NFSD needs to get the test kernel too.

I tried adding nfsd to [baseline], but then "make nfstest" set up
the test programs on the nfsd host as well as the clients.

Fixes: 434df33309e7 ("workflows: linux: remove explicit inventory in ansible wrappers")
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>