terraform/gce: Remove the kconfigs/images subdirectory
Instead of sourcing files in a subdirectory, move the content of
those files right into Kconfig.image. This will make it easier
to examine the differences between the current menu content and
the dynamically-generated content.

The consolidation is bug-for-bug compatible with the original
files, including missing default selections in some choice blocks.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>