terraform/gce: Remove the kconfigs/zones subdirectory
Instead of sourcing files in a subdirectory, move the content of
those files right into Kconfig.location. This will make it easier
to examine the differences between the current menu content and
the dynamically-generated content.

The pattern of multiple TERRAFORM_GCE_ZONE declarations is valid
Kconfig since each is in a mutually exclusive if block (only one
region can be active at a time).

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>